+++ NEWS.md
@@ -56,6 +56,7 @@ New library features
 --------------------
 * `replace(string, pattern...)` now supports an optional `IO` argument to
   write the output to a stream rather than returning a string ([#48625]).
+* `sizehint!(s, n)` now supports an optional `shrink` argument to disable shrinking ([#51929]).
 
 Standard library changes
 ------------------------